Create Data Acquisition from Stack Exchange
The goal of Task 1 was to acquire data from Stack Exchange of the top 200,000 posts by View Count. Firstly, we identified what the lowest value of View Count there was in the top 200,000 posts. We then ran a query as demonstrated below and identified there were 200,002 posts with a View Count greater than 41231, including 199,999 posts with a View Count greater than 41232. This meant it was impossible to collect exactly the top 200,000 posts as there were three posts that had a View Count of 41232.

- Count Query #1: select count(*) from posts where posts.ViewCount > 41231
- Count Query #2: select count(*) from posts where posts.ViewCount > 41232

Due to the inherent functionality restrictions of Stack Exchange Data Explorer, an individual can only download a maximum of 50,000 records at a time. This meant we had to run 5 queries as can be seen below from the Stack Exchange site to get the 200,002 posts identified as above.

- Query #1: select top 50000* from posts where posts.ViewCount > 41231 order by posts.ViewCount DESC
- Query #2: select top 50000 * from posts where posts.ViewCount <= 127078 order by posts.ViewCount DESC
- Query #3:  select top 50000 * from posts where posts.ViewCount <= 74420 order by posts.ViewCount DESC
- Query #4: select top 50000 * from posts where posts.ViewCount <= 53077 order by posts.ViewCount DESC
- Query #5: select top 50000 * from posts where posts.ViewCount <= 41234 and posts.ViewCount > 41231 order by posts.ViewCount DESC

Please note the upper bound also has an equal to sign (<=) for Queries #2, #3, #4 and #5. This is to include the lowest view count value in the previous query as there may be posts with the same view count value, but it was not included in the dataset due to the limit in selecting the top 50,000 posts. Otherwise, using a less than sign (<) may exclude posts that are in the top 200,002 posts by View Count. As a result, there may be duplicates in the total dataset, however these will be identified and removed in the data processing, cleaning and integration section of this assignment and verified with a count of 200,002 for completeness and accuracy.